+++ src/components/CardHeaderIn.tsx
@@ -27,7 +27,7 @@ const CardHeaderIn = ({
   return (
     <Card style={style}>
       <View style={styles.cardHeaderView}>
-        <Text style={styles.cardHeaderText} colorVariant={'text2'}>
+        <Text style={styles.cardHeaderText}>
           {topText}
         </Text>
       </View>
