+++ src/screens/etis/orders/OrderModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Modal, View } from 'react-native';
+import { Modal, StyleSheet, View } from 'react-native';
 import AutoHeightWebView from 'react-native-autoheight-webview';
 import ClickableText from '~/components/ClickableText';
 import { useGlobalStyles } from '~/hooks';
@@ -19,33 +19,30 @@ export default function OrderModal({ html, closeModal }) {
 
   return (
     <Modal transparent onRequestClose={closeModal}>
-      <View
-        style={[
-          {
-            alignItems: 'center',
-            justifyContent: 'center',
-            flex: 1,
-            marginVertical: '20%',
-            marginHorizontal: '2%',
-          },
-          globalStyles.block,
-          globalStyles.border,
-          globalStyles.shadow,
-        ]}
-      >
+      <View style={[styles.container, globalStyles.card, globalStyles.border]}>
         <AutoHeightWebView
           source={{ html }}
           scalesPageToFit
           style={{ marginHorizontal: '20%' }}
-          customStyle={getStyles(globalStyles.fontColorForBlock.color)}
+          customStyle={getStyles(globalStyles.textColor.color)}
         />
         <ClickableText
           text={'Закрыть'}
           onPress={closeModal}
-          textStyle={[fontSize.large, globalStyles.fontColorForBlock]}
+          textStyle={[fontSize.large, globalStyles.textColor]}
           viewStyle={{ marginBottom: '2%' }}
         />
       </View>
     </Modal>
   );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    flex: 1,
+    marginVertical: '20%',
+    marginHorizontal: '2%',
+  },
+});
